docs: add placeholder for other package managers
The lithology-patterns isn't npm only package as it just provides the patterns. For the wellx project purpose we bundled these svgs with npm to distribute them as a package. However, npm isn't the only package manager out there and others are welcome to us the same library to bundle theirs.